pyproject.toml
[tool.poetry]
name = "medical-search"
version = "0.1.0"
description = ""
authors = ["chun hsuan <hsuan069317@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.58.1"
langchain-openai = "^0.2.13"
langchain = "^0.3.13"
langchain-community = "^0.3.13"
langchain-core = "^0.3.27"
chromadb = "^0.5.23"
streamlit = "^1.41.1"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
aiohttp = "^3.11.11"
watchdog = "^6.0.0"
python-dotenv = "^1.0.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
